Separating `ActiveFilterItem` component
